// apps/web/src/lib/api/generated/model/insurerOut.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * BrokerOS API
 * OpenAPI spec version: 1.0.0
 */
import type { InsurerOutBankDetails } from './insurerOutBankDetails';

export interface InsurerOut {
  bank_details: InsurerOutBankDetails;
  email: string | null;
  id: string;
  is_active: boolean;
  mpesa_paybill: string | null;
  name: string;
  notes: string | null;
  payment_account_hint: string | null;
  phone: string | null;
  short_name: string | null;
  version: number;
}